Created/added test file for 4 digit (speedometer) display
Got the 4 digit display working. The trick was that the Cursor for entering the digits is saved on the atmel processor embedded on the device, so it is a saved state, even if you reset the main arduino. To get around this, I set the cursor to the left-most digit on setup.