Rewrite Elcano_Serial for Arduinos
This rewrites the Elcano_Serial library more in the style of a
lexer/parser, as it is now completely bufferless (outside of the API)
and more stable than the previous version, according to Chase. While not
perfect (that would require a state machine, which is still being mapped
out), this will be neccessary for the presentation next week.