Asyncronous Non-blocking Parsing
Previously, the parser would read from an internal buffer, which would
be built up over a period of time. This would have worked for a mere
demo, but in practice would require hard blocks for somewhat arbitrary
units of time, depending on the amount of work being done on the other
end of the line.

This fixes this issue, by restructuring the parser to use a state
machine. With this, the parser calls an update method in the "void
loop();" method, and if it returns 1, we have a complete package to work
with. This gets rid of any delays, locks, or blocks, and theorettically
even allows us to read from multiple connections at once.